create actor queries and address resolution sub module in state manager
